Rename package structure from atlassian.connect.jwt.scala to atlassian.jwt to match the project name
