+++ modules/api/src/main/scala/io/toolsplus/atlassian/jwt/api/Predef.scala
@@ -1,0 +1,12 @@
+package io.toolsplus.atlassian.jwt.api
+
+object Predef {
+
+  /** Type indicating a raw JWT string.
+    *
+    * Prefer this type anywhere there is a raw JWT string over a plain
+    * string type.
+    */
+  type RawJwt = String
+
+}
