+++ modules/core/src/test/scala/io/toolsplus/atlassian/jwt/util/CommonGen.scala
@@ -1,0 +1,12 @@
+package io.toolsplus.atlassian.jwt.util
+
+import org.scalacheck.Gen
+import org.scalacheck.Gen._
+
+trait CommonGen {
+
+  def alphaStr(n: Int): Gen[String] = listOf(n, alphaChar).map(_.mkString)
+
+}
+
+object CommonGen extends CommonGen
