--- modules/generators/src/main/scala/io/toolsplus/atlassian/connect/jwt/scala/generators/core/JwtGen.scala
@@ -1,3 +1,0 @@
-package io.toolsplus.atlassian.connect.jwt.scala.generators.core
-
-trait JwtGen extends CanonicalHttpRequestGen
