--- modules/generators/src/main/scala/io/toolsplus/atlassian/connect/jwt/scala/generators/util/JwtTestHelper.scala
@@ -1,7 +1,0 @@
-package io.toolsplus.atlassian.connect.jwt.scala.generators.util
-
-object JwtTestHelper {
-
-  val defaultSigningSecret: String = "EVqrJGaYbf=EGVwg)aYAxXJqm2zjfab6"
-
-}
