+++ modules/generators/src/main/scala/io/toolsplus/atlassian/jwt/generators/core/JwtGen.scala
@@ -1,0 +1,3 @@
+package io.toolsplus.atlassian.jwt.generators.core
+
+trait JwtGen extends CanonicalHttpRequestGen
