+++ modules/generators/src/main/scala/io/toolsplus/atlassian/jwt/generators/util/JwtTestHelper.scala
@@ -1,0 +1,7 @@
+package io.toolsplus.atlassian.jwt.generators.util
+
+object JwtTestHelper {
+
+  val defaultSigningSecret: String = "EVqrJGaYbf=EGVwg)aYAxXJqm2zjfab6"
+
+}
